% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportsfeatures-package.R
\docType{package}
\name{sportsfeatures-package}
\alias{sportsfeatures}
\alias{sportsfeatures-package}
\title{sportsfeatures package documentation}
\description{
A synthetic, longitudinal athletic dataset generated through a transparent, rule-based simulation engine. Captures individual activity sessions across multiple athletes, environmental conditions, and physiological responses. Specifically designed as an alternative to legacy teaching datasets by introducing realistic hierarchical repeated measures, complex two-way covariate interactions, and a deliberate Missing Not At Random (MNAR) tracking mechanism suitable for advanced imputation workflows. Methodologies implemented are based on van Buuren (2018) \doi{10.1201/9780429492259} and Bates et al. (2015) \doi{10.18637/jss.v067.i01}.
}
\author{
\strong{Maintainer}: Mohammad Abbas \email{ma.abbas3107@gmail.com}

Authors:
\itemize{
  \item Mohammad Abbas \email{ma.abbas3107@gmail.com}
}

}
\keyword{internal}
